+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # foundry-vtt-expiry
 
+## 1.0.2
+- Configuration dialog improvements
+- Basic logic implementation for item name changes
+
 ## 1.0.1
+- Initial POC code